Restructuring edge cases test classes
